flash sales
installed axios to fetch data from api for slash sales.  fetched data from te api. stylying is not finished. timer needs to be fixed. needs to be worked in real time. also one feature needs to added - when the countdown the time is up the discount should dissapear and those products prices should go back to the original one.